+++ windows/cloud.ps1
@@ -26,9 +26,18 @@ function Get-PowerShellScript ($script_name, $script_url) {
 }
 
 # Download utils PowerShell script with helper functions and import it, and download cloud-1.ps1
-$utils_script_name = "C:\utils.psm1"
-$utils_script_url = "https://fractal-cloud-setup-s3bucket.s3.amazonaws.com/utils.psm1"
-Get-PowerShellScript $utils_script_name $utils_script_url
+if (-not (Test-Path env:LOCAL)) { $env:LOCAL = 'no' }
+If ($env:LOCAL  -eq 'no')  {
+    Write-Output "Downloading utils"
+    $utils_script_name = "C:\utils.psm1"
+    $utils_script_url = "https://fractal-cloud-setup-s3bucket.s3.amazonaws.com/utils.psm1"
+    Get-PowerShellScript $utils_script_name $utils_script_url
+} Else {
+    Write-Output "Running Local: LOCAL=$env:LOCAL"
+    Get-Location
+    $utils_script_name = "./utils.psm1"
+}
+
 Import-Module "$utils_script_name"
 
 # Make sure we're in the proper directory (mostly for running from webservers) and define credentials for user
@@ -147,5 +156,8 @@ if ($productivity_install) {
 
 # Clean PowerShell install script and restart
 Write-Output "Cleaning up Utils script script"
-Remove-Item -Path $utils_script_name -Confirm:$false
-Restart-Computer -Force
+Remove-Item -Path $utils_script_name -Confirm:$false -ErrorAction SilentlyContinue
+
+If ($LOCAL  -eq 'no')  {
+    Restart-Computer -Force
+}
